requirements.txt
pdfplumber
langchain
openai
faiss-cpu
python-docx
reportlab
Pillow
PyYAML
fastapi
uvicorn
PyPDF2
streamlit
requests
